unscanny

Trait Pattern

pub trait Pattern<T>: Sealed<T> { }

Something a string can start with.

This is implemented for:

Type	Example

char	scanner.at('a')

&str	scanner.at("hello")

[char; N], &[char]	scanner.at(['a', 'b', 'c'])

FnMut(char) -> bool	scanner.at(char::is_alphabetic)

FnMut(&char) -> bool	scanner.at(char::is_ascii)

As you might have noticed, this closely mirrors the Pattern trait from the standard library. Unfortunately, this trait is unstable, so we can’t use it in the scanner’s method signatures. Furthermore, it doesn’t support passing &char functions which is quite useful because some char methods take self by reference.